basic_and_jwt/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.1.9
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.6.3
